+++ src/Hartigan.hpp
@@ -18,9 +18,9 @@ namespace gmum {
   public:
     Hartigan(bool logNrOfClusters, bool logEnergy);
     TotalResult loop(arma::mat &points, std::vector<unsigned int> &assignment,
-		       float killThreshold, std::vector<boost::shared_ptr<Cluster> > &clusters);
+		       double killThreshold, std::vector<boost::shared_ptr<Cluster> > &clusters);
     SingleResult singleLoop(arma::mat &points, std::vector<unsigned int> &assignment, 
-			      float killThreshold, std::vector<boost::shared_ptr<Cluster> > &clusters);
+			      double killThreshold, std::vector<boost::shared_ptr<Cluster> > &clusters);
   };
 
 }
